# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QNAP File Management API",
        "description": "This API enables file upload and download operations on a QNAP server, facilitating data management. It provides an example of how to interact with the QNAP server using PHP, allowing developers to easily integrate file management features into their applications.",
        "contact": {
            "name": "Deivid Medina",
            "email": "-"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/dnda/files/upload": {
            "post": {
                "tags": [
                    "File"
                ],
                "summary": "Upload a file to QNAP",
                "description": "\n     *  Example request:\n     *      POST api/v1/dnda/files/upload?sid={value}&folderPath={value}\n     *          \n     *      form-data\n     *          KEY: file\n     *          VALUE: fileName.xxxx         \n     * \n     *  Example of a successful response:\n     *      {\n     *          'status': 1,\n     *          'size': '6274227',\n     *          'name': 'example.pdf'\n     *      }\n     * \n     *  Example of an unsuccessful response:\n     *  Refer to the documentation: https://download.qnap.com/dev/QNAP_QTS_File_Station_API_v5.pdf\n     *      {\n     *          'version': '5.5.5',\n     *          'build': '20240817',\n     *          'status': 9,\n     *          'success': 'true'\n     *      }\n     *      ",
                "operationId": "b21678ab2dbcdbcc7b940ea85d028573",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "sid": {
                                        "description": "User session ID",
                                        "type": "string"
                                    },
                                    "folderPath": {
                                        "description": "Folder path where the file will be uploaded",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "sid": "12345",
                                    "folderPath": "/my/folder",
                                    "file": "my_file.txt"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "1."
                                        },
                                        "size": {
                                            "type": "string",
                                            "example": "4514151."
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "fileName.xxxx"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable content, documentation: https://download.qnap.com/dev/QNAP_QTS_File_Station_API_v5.pdf",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "string",
                                            "example": "5.5.5"
                                        },
                                        "build": {
                                            "type": "string",
                                            "example": "20240817"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "9"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dnda/files/download": {
            "get": {
                "tags": [
                    "File"
                ],
                "summary": "Download a file from the specified folder",
                "description": "\n     *  Example request:\n     *      POST api/v1/dnda/files/download?sid={value}&folderPath={value}&fileName={value}\n     *               \n     *  Example of a successful response:\n     *      File downloaded\n     * \n     *  Example of an unsuccessful response:\n     *  Refer to the documentation: https://download.qnap.com/dev/QNAP_QTS_File_Station_API_v5.pdf\n     *      404 not found\n     *      ",
                "operationId": "9b4ccd9fdc8c4a1c5b42ae52ceac524f",
                "parameters": [
                    {
                        "name": "sid",
                        "in": "query",
                        "description": "User session ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderPath",
                        "in": "query",
                        "description": "Path where the file to be downloaded is located",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fileName",
                        "in": "query",
                        "description": "Name of the file, including extension, that you want to download",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully"
                    },
                    "404": {
                        "description": "Content not found: https://download.qnap.com/dev/QNAP_QTS_File_Station_API_v5.pdf"
                    }
                }
            }
        },
        "/api/v1/dnda/folder/exists": {
            "get": {
                "tags": [
                    "Folder"
                ],
                "summary": "Validate if a folder exists in a specific location",
                "description": "\n     *  Example request:\n     *      POST api/v1/dnda/folder/exists?sid={value}&folderPath={value}&folderName={value}\n     *               \n     *  Example of a successful response:\n     *      True - If the folder already exists in the location\n     * \n     *  Example of an unsuccessful response:\n     *      False - If the folder does not exist in the location\n     *      ",
                "operationId": "652710428b5c7c07513c047162f07eac",
                "parameters": [
                    {
                        "name": "sid",
                        "in": "query",
                        "description": "User session ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderPath",
                        "in": "query",
                        "description": "Location of the folder to validate",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderName",
                        "in": "query",
                        "description": "Name of the folder to validate",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "True if the folder exists in the location.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "true": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "False if the folder does not exist in the location.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "false": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dnda/folder/create": {
            "post": {
                "tags": [
                    "Folder"
                ],
                "summary": "Create a folder in a specified location",
                "description": "\n     *  Example request:\n     *      POST api/v1/dnda/folder/create?sid={value}&folderPath={value}&folderName={value}\n     *               \n     *  Example of a successful response:\n     *      {\n     *          'version': '5.5.5',\n     *          'build': '20240817',\n     *          'status': 1,\n     *          'success': 'true'\n     *      }\n     * \n     *  Example of an unsuccessful response:\n     *  Refer to the documentation: https://download.qnap.com/dev/QNAP_QTS_File_Station_API_v5.pdf\n     *      {\n     *          'version': '5.5.5',\n     *          'build': '20240817',\n     *          'status': 33,\n     *          'success': 'true'\n     *      }\n     * ",
                "operationId": "5634b28b0c1943447fd254d2b8cb9ee9",
                "parameters": [
                    {
                        "name": "sid",
                        "in": "query",
                        "description": "User session ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderPath",
                        "in": "query",
                        "description": "Location where the folder will be created",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderName",
                        "in": "query",
                        "description": "Name of the folder to be created",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If the creation was successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "string",
                                            "example": "5.5.5"
                                        },
                                        "build": {
                                            "type": "string",
                                            "example": "20240817"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dnda/session/login": {
            "post": {
                "tags": [
                    "Session"
                ],
                "summary": "Obtain a Sid token",
                "description": "\n     *  Example request:\n     *      POST api/v1/dnda/session/login?username={value}&password={value}\n     *               \n     *  Example of a successful response:\n     *      {Sid}\n     * \n     *  Example of an incorrect response:\n     *  Refer to the documentation: https://download.qnap.com/dev/QNAP_QTS_File_Station_API_v5.pdf\n     *      403 Forbidden\n     *      ",
                "operationId": "e1d9d60c9d395ce2a3b410f75d41489d",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "QNAP user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Base64 encoded password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sid Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "dasjdhsano": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "File",
            "description": "File"
        },
        {
            "name": "Folder",
            "description": "Folder"
        },
        {
            "name": "Session",
            "description": "Session"
        }
    ]
}